Add files via upload
Using the template wordcount.asm as a starting point, you are to write a MIPS program that counts the number of occurrences of a word in a text block entered using the keyboard. The user enters a text block (no more than 600 characters) in the console and then asks for a word to search for. Note that a word means any continuous set of characters before the next space. The program should search the text block and display the number of occurrences of the word entered. The program should print the text and the word entered by the keyboard. The end of the text block is reached when the user presses the <Enter> key. After the program finishes it should provide an option of redoing the task or exiting the program.
You should implement the program using memory-mapped I/O, i.e., you are not allowed to use the syscall. A sample output would be like this:
    Word count
    Enter the text segment:
        Roses are red, the sky is blue and firetrucks are red.
    Enter the search word:
        red
    The word ’red’ occurred 2 time(s).
    press ’e’ to enter another segment of text or ’q’ to quit.
1
Your code should handle cases where the search word is not in the text block. For example:
    Word count
    Enter the text segment:
        I love assembly language. Language is my thing!
    Enter the search word:
        computer
    The word ’computer’ occurred 0 time(s).
    press ’e’ to enter another segment of text or ’q’ to quit.
Note that for this assignment you need to use the “Keyboard and Display MMIO Sim- ulator” in MARS by connecting it to MIPS. This can be found under Tools->Keyboard and Display MMIO Simulator. Click Connect to MIPS. In the main MARS window, after assembling your code, select the 0xffff0000 (MMIO) portion of the data segment. This is the memory mapped IO area. Click Reset in the MMIO simulator window.