An 8-bit shift register
Design an 8-bit shift register with output bits A7A6A5A4A3A2A1A0. Each bit Ai is repre- sented by the Q output of a rising edge triggered D Flip-flop, which is available in logisim- evolution. The circuit has a clock and it also has the following two selector inputs: S0,S1. The roles of the selector inputs are described in the table below. The timing is such that a hold, shift right, shift left or circular shift right, is controlled by the clock input.
The bit that is shifted out, i.e., when you do a shift right or a shift left, should be stored in an additional output called Bout. Correspondingly, an input bit, called Bin should be brought in to the space created by the open position, when you carry out a shift operation. A template file has been provided for you called shift.circ, which you can start with and then modify. Explain your design in words in a separate text file, which you will also upload, and then submit the circuit (shift.circ) implementation in logisim-evolution. Note that the TA’s will download and test your circuit design.

S1   S0  function
0      0    no shift -hold
0      1     shift right
1      0     shift left
1      1      circular shift right