String capitalization
You are to write a MIPS program that takes as input a string of text and then does the following.
• Prints the input string
• Capitalizes and prints the capitalized string.
You can assume that the string input is comprised of words consisting of keyboard alpha- betic characters, with each word separated by a single “space” character, and with each word beginning with an alphabetical character. The program should be able to handle strings up to 127 characters in length. No characters other than lowercase letters should be modified. In other words, the input string may have some of its letters already capitalized. Use properties of the ASCII code, and in particular the manner in which characters are represented, to solve this problem!
As a working example, given the input string "this is A StrinG", the printed capitalized output should be "THIS IS A STRING". A template file capitalize.asm has been provided for you. In this file, you will note that there is an input string that is hard-coded and there is place for an output string. The TA’s might change the contents of the input string while testing your code. You should use the space allocated for the output string to write in the content of your capitalized output. Then, it should be a simple matter to print it.